move all typed ast enums to ast module along with it's methods
